fix: Cannot refer toplevel constant from a class
ffe30d9 changed convert_program to process toplevel_defs before
main_exprs (so that you can call auto-defined getters in main_exprs).
However, for example, this program could not be compiled due to the
change:

  C = 1
  class A; def foo; C; end; end

In this program constant C defined in main_exprs is referred from
class A, which is in the toplevel_defs.

To solve this, this commit introduces ast::TopLevelItem and process
each definitions and expressions in the lexical order.