fix: Add HirLambdaExpr.ret_ty
If a lambda uses `break`, its return type must be Void. However you
cannot tell the return type of block until it is passed to a method.
So I'm using `std::mem::swap` to set `ret_ty` (pretty adhoc, but I
think this will be deleted when type inference is implemented)