Allow methods of classes not in skc_corelib to be defined in Rust
Currently rustlib_methods assumes all the classes in skc_rustlib
are defined in skc_corelib. For example, if you want to add
builtin/time.sk and define its method in skc_rustlib, you need to
edit skc_corelib too.

This is not only redundant but also could be a problem when
implementing library system (like RubyGems).